Create code 2
This function takes two parameters: temperature and humidity. It uses if-else statements to determine the probability of precipitation based on the given temperature and humidity. The probability is calculated by multiplying the humidity by a factor that depends on the temperature.

Please note that this function is just an example and it does not reflect the real-world physics, and the real-world data would be more complicated to calculate, and other factors such as pressure, wind, and front systems would also have to be considered.